# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "srvcs-outofrange",
    "description": "",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/": {
      "get": {
        "tags": [],
        "summary": "`GET /` — service identity (srvcs service standard).",
        "operationId": "index",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Info"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [],
        "summary": "`POST /` — compute `outofrange(value, lo, hi)` by composing two primitives.",
        "description": "This service owns the *control flow* but delegates the actual logic to its\ndependencies, exactly as specified:\n\n1. ask `srvcs-between` for `b = between(value, lo, hi)`;\n2. ask `srvcs-not` for `result = not(b)` — i.e. value is outside `[lo, hi]`.\n\nValidation propagates from the dependencies: this orchestrator never calls\n`srvcs-isnumber` directly. If a dependency is unreachable it reports itself\ndegraded (`503`); if a dependency rejects the input it forwards the `422`.",
        "operationId": "evaluate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutOfRangeResponse"
                }
              }
            }
          },
          "422": {
            "description": "a dependency rejected the input (forwarded)"
          },
          "500": {
            "description": "a dependency returned a malformed result"
          },
          "503": {
            "description": "a dependency is unavailable"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "EvalRequest": {
        "type": "object",
        "required": [
          "value",
          "lo",
          "hi"
        ],
        "properties": {
          "hi": {
            "type": "number",
            "format": "double"
          },
          "lo": {
            "type": "number",
            "format": "double"
          },
          "value": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "Info": {
        "type": "object",
        "required": [
          "service",
          "concern",
          "depends_on"
        ],
        "properties": {
          "concern": {
            "type": "string"
          },
          "depends_on": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "service": {
            "type": "string"
          }
        }
      },
      "OutOfRangeResponse": {
        "type": "object",
        "required": [
          "value",
          "lo",
          "hi",
          "result"
        ],
        "properties": {
          "hi": {
            "type": "number",
            "format": "double"
          },
          "lo": {
            "type": "number",
            "format": "double"
          },
          "result": {
            "type": "boolean"
          },
          "value": {
            "type": "number",
            "format": "double"
          }
        }
      }
    }
  }
}